refactor(test): split metadata-references by family, off the 100KB ceiling (#814)
`test/metadata-references.test.ts` had reached 99,872 bytes against the
`MAX_FILE_BYTES = 100 * 1024` cap in `scripts/check-source-hygiene.mjs` —
97.5% of the quota, about one commented guard of headroom. #815 already paid
for that once: it set out to invert a single bulk-dispatch guard, blew the
cap, and had to design an unplanned split mid-PR before its real change could
be reviewed. Every later PR touching these guards was queued to repeat it.

The 22 describe blocks are re-homed by the metadata surface each resolves
against:

  metadata-references  pages, forms, and the cross-surface references
                       (formula predicates, flow conditions, objectOverride,
                       dashboard global filters, App.defaultAgent)   24 tests
  view-references      view fields, sorts, filter template tokens, row
                       colours, kanban groups, stage enumerations, list-view
                       reachability                                  16 tests
  action-references    navigation, dashboard actions and routes, list-level
                       rowAction / bulkAction, dashboard date ranges 11 tests
  i18n-references      action labels, select fields and options, locale
                       completeness on every authored surface        19 tests

Shared derivations (objects, views, walk, the flattened locale packs, the
platform-object allowlist) move to `test/helpers/metadata-fixtures.ts` — one
definition instead of four. `metadata-references.test.ts` keeps a map at the
top naming which family went where, so the existing comments in `src/` and in
`bulk-action-dispatch.test.ts` that point at this file still land on their
guard in one hop.

Semantics are unchanged and were verified mechanically: every describe body
was diffed against the original and is byte-identical, except five comment
sentences the split itself falsified ("the navigation guard in this file",
which now names `test/action-references.test.ts`). Reversing those five
restores the original bytes exactly. 24 + 16 + 11 + 19 = 70 tests, the same 70
the single file ran; the full suite is 1597 passed | 1 skipped on both sides.

Largest resulting file is 33,451 bytes — 67% below the ceiling.
