+++ test/metadata-references.test.ts
@@ -1,10 +1,20 @@
 // Copyright (c) 2025 ObjectStack. Licensed under the Apache-2.0 license.
 
 import { describe, it, expect } from 'vitest';
-import { isDateMacroToken } from '@objectstack/spec/data';
 import { AgentSchema } from '@objectstack/spec/ai';
 import stack from '../objectstack.config';
-import { OPPORTUNITY_STAGE_OPTIONS } from '../src/objects/_picklists';
+import {
+  type AnyRec,
+  objects,
+  pages,
+  views,
+  profiles,
+  objectNames,
+  profileNames,
+  PLATFORM_OBJECTS,
+  fieldsOf,
+  walk,
+} from './helpers/metadata-fixtures';
 
 /**
  * Dangling-reference guards for UI metadata.
@@ -18,86 +28,42 @@ import { OPPORTUNITY_STAGE_OPTIONS } from '../src/objects/_picklists';
  *
  * These tests resolve every UI reference against the objects/profiles the app
  * really defines, so the next bad name fails in CI instead of in a demo.
- */
-
-type AnyRec = Record<string, any>;
-const objects: AnyRec[] = (stack as any).objects ?? [];
-const pages: AnyRec[] = (stack as any).pages ?? [];
-const views: AnyRec[] = (stack as any).views ?? [];
-const profiles: AnyRec[] = (stack as any).permissions ?? [];
-const stackActions: AnyRec[] = (stack as any).actions ?? [];
-
-const objectNames = new Set(objects.map((o) => o.name));
-const profileNames = new Set(profiles.map((p) => p.name));
-
-/**
- * Locale packs, flattened to `[locale, pack]` pairs.
  *
- * `stack.translations` holds ONE `TranslationBundle` keyed by locale
- * (`{ en: {...}, 'zh-CN': {...} }`) — NOT a list of per-locale records. A
- * `translations.find(t => t.locale === 'zh-CN')` therefore matches nothing and
- * silently turns its test into a no-op, which is how the navigation guard
- * below spent its life passing without asserting anything.
+ * ---
+ *
+ * SPLIT BY FAMILY (#814) — where the other guards went.
+ *
+ * This file reached 99,872 bytes against the 100KB ceiling in
+ * `scripts/check-source-hygiene.mjs`, so the next PR to add a guard would have
+ * been failed by `pnpm hygiene` before review (#815 had already been forced
+ * into one unplanned split for that reason). It now keeps the PAGE, FORM and
+ * cross-surface guards; three siblings own the rest, and each holds the whole
+ * family so a reader lands on the guard in one hop:
+ *
+ *   `test/view-references.test.ts`
+ *     view field references resolve · priority queues sort by urgency, not
+ *     alphabetically · filter template tokens are resolvable · row colors and
+ *     kanban groups key off real option values · every canonical opportunity
+ *     stage reaches the UI that enumerates stages · every named list view is
+ *     reachable
+ *
+ *   `test/action-references.test.ts`
+ *     navigation reaches everything the app ships · dashboard actions land on
+ *     real routes · list-level action references resolve · dashboard date
+ *     ranges window a field the query layer can actually compare
+ *
+ *   `test/i18n-references.test.ts`
+ *     picklist values never reach the UI unresolved · action labels are
+ *     translated in every locale · select fields are translated in every
+ *     locale · every locale is complete on every authored surface
+ *
+ *   `test/bulk-action-dispatch.test.ts` (split earlier, by #815)
+ *     the `bulkActionDefs` / `execution: 'aggregate'` dispatch contract
+ *
+ * The derivations all four share live in `test/helpers/metadata-fixtures.ts`.
+ * The split moved text only: no assertion, helper or fixture changed, and the
+ * suite runs the same 70 tests it ran before.
  */
-const localePacks: [string, AnyRec][] = ((stack as any).translations ?? []).flatMap(
-  (bundle: AnyRec) => Object.entries(bundle) as [string, AnyRec][],
-);
-const packFor = (locale: string): AnyRec | undefined =>
-  localePacks.find(([name]) => name === locale)?.[1];
-
-/**
- * Audit columns the platform adds to every object. They are real at runtime
- * (`?sort=created_at desc` works) but never appear in the authored `fields`
- * map, so a reference to one is legitimate.
- */
-const SYSTEM_FIELDS = ['id', 'created_at', 'updated_at', 'created_by', 'updated_by'];
-
-/**
- * System objects the installed platform plugins actually register, verified
- * against the 16.1.0 bundles in node_modules:
- *   - @objectstack/platform-objects — sys_user, sys_email, sys_file, … (the
- *     large core set; only the ones app metadata references are listed here)
- *   - @objectstack/plugin-approvals — sys_approval, sys_approval_request,
- *     sys_approval_action, … (note: there is NO `sys_approval_process`)
- *   - the activity/comment timeline objects used by `record:activity` and the
- *     log_call / log_meeting actions
- * A reference to a `sys_*` name outside this set matches nothing at runtime.
- */
-const PLATFORM_OBJECTS = new Set([
-  'sys_user',
-  'sys_organization',
-  'sys_team',
-  'sys_email',
-  'sys_email_template',
-  'sys_file',
-  'sys_attachment',
-  'sys_notification',
-  'sys_inbox_message',
-  'sys_activity',
-  'sys_comment',
-  'sys_approval',
-  'sys_approval_request',
-  'sys_approval_action',
-  'sys_approval_approver',
-  'sys_approval_delegation',
-]);
-
-const fieldsOf = (obj: string) => [
-  ...Object.keys(objects.find((o) => o.name === obj)?.fields ?? {}),
-  ...SYSTEM_FIELDS,
-];
-
-/** Walk an arbitrary metadata tree, yielding every node that has a `type`. */
-function* walk(node: unknown): Generator<AnyRec> {
-  if (Array.isArray(node)) {
-    for (const item of node) yield* walk(item);
-    return;
-  }
-  if (!node || typeof node !== 'object') return;
-  const rec = node as AnyRec;
-  if (typeof rec.type === 'string') yield rec;
-  for (const value of Object.values(rec)) yield* walk(value);
-}
 
 describe('page component references resolve', () => {
   const components = pages.flatMap((p) => [...walk(p.regions), ...walk(p.slots)]);
@@ -290,56 +256,6 @@ describe('page component references resolve', () => {
   });
 });
 
-describe('view field references resolve', () => {
-  /** Views are keyed by object; `defineView` output carries the object on its data provider. */
-  const viewObjectOf = (v: AnyRec): string | undefined =>
-    v.list?.data?.object ?? v.form?.data?.object ?? v.object;
-
-  it('every form section field is a real field on the view object', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const objectName = viewObjectOf(v);
-      if (!objectName || !objectNames.has(objectName)) continue;
-      const known = fieldsOf(objectName);
-      // The default `form` plus every named form under `formViews`. (The
-      // container key is `formViews` — iterating a non-existent `forms` key
-      // silently skipped every named form view.)
-      const forms = [v.form, ...Object.values(v.formViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const form of forms) {
-        for (const section of form.sections ?? []) {
-          for (const f of section.fields ?? []) {
-            const name = typeof f === 'string' ? f : f?.field;
-            if (name && !known.includes(name)) {
-              bad.push(`${objectName} form "${form.name ?? 'default'}": no field "${name}"`);
-            }
-          }
-        }
-      }
-    }
-    expect(bad, `dangling form fields:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every list sort targets a real field', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const objectName = viewObjectOf(v);
-      if (!objectName || !objectNames.has(objectName)) continue;
-      const known = fieldsOf(objectName);
-      // The container key is `listViews` — iterating a non-existent `views`
-      // key silently skipped every named list view.
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        for (const s of list.sort ?? []) {
-          if (s.field && !known.includes(s.field)) {
-            bad.push(`${objectName} view "${list.name ?? 'default'}": sorts on missing "${s.field}"`);
-          }
-        }
-      }
-    }
-    expect(bad, `dangling sort fields:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
 describe('formula fields are never used as query predicates', () => {
   /**
    * A `formula` field is not a column. The engine evaluates it in JS over the
@@ -459,361 +375,6 @@ describe('flow conditions reach the CEL engine', () => {
   });
 });
 
-describe('priority queues sort by urgency, not alphabetically', () => {
-  /**
-   * `priority desc` on the select itself compares the raw option strings, so
-   * cases came back medium > low > high > critical — the exact inversion of
-   * urgency. Queue views must sort on the materialised `priority_rank`.
-   */
-  const rankedObjects = ['crm_case', 'crm_task'];
-
-  it.each(rankedObjects)('%s defines a numeric priority_rank', (objectName) => {
-    const field = objects.find((o) => o.name === objectName)?.fields?.priority_rank;
-    expect(field, `${objectName}.priority_rank missing`).toBeTruthy();
-    expect(field.type).toBe('number');
-  });
-
-  it('no view sorts on the raw priority select', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        for (const s of list.sort ?? []) {
-          if (s.field === 'priority') {
-            bad.push(`view "${list.name ?? list.label ?? 'default'}" sorts on raw priority`);
-          }
-        }
-      }
-    }
-    expect(bad, `${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-describe('navigation reaches everything the app ships', () => {
-  const apps: AnyRec[] = (stack as any).apps ?? [];
-  const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
-  const reports: AnyRec[] = (stack as any).reports ?? [];
-
-  /** Flatten the (possibly nested) navigation tree. */
-  const navNodes = (app: AnyRec): AnyRec[] =>
-    (app.navigation ?? []).flatMap(function walk(n: AnyRec): AnyRec[] {
-      return [n, ...(n.children ?? []).flatMap(walk)];
-    });
-  const allNodes = apps.flatMap(navNodes);
-
-  it('every nav entry points at something that exists', () => {
-    const dashboardNames = new Set(dashboards.map((d) => d.name));
-    const reportNames = new Set(reports.map((r) => r.name));
-    const bad: string[] = [];
-    for (const n of allNodes) {
-      // `sys_*` are platform objects the app legitimately links to, but only
-      // ones an installed plugin actually registers. A `requiresObject` guard
-      // hides the item gracefully when a plugin is absent — it does not
-      // legitimise pointing at an object NO installed plugin ever registers
-      // (`sys_approval_process` was permanent dead weight: the guard hid it on
-      // every install, forever).
-      if (n.objectName && n.objectName.startsWith('sys_')) {
-        if (!PLATFORM_OBJECTS.has(n.objectName)) {
-          bad.push(`${n.id}: system object "${n.objectName}" is not registered by any installed plugin`);
-        }
-      } else if (n.objectName && !objectNames.has(n.objectName)) {
-        bad.push(`${n.id}: object "${n.objectName}" is not defined`);
-      }
-      if (n.dashboardName && !dashboardNames.has(n.dashboardName)) {
-        bad.push(`${n.id}: dashboard "${n.dashboardName}" is not defined`);
-      }
-      if (n.reportName && !reportNames.has(n.reportName)) {
-        bad.push(`${n.id}: report "${n.reportName}" is not defined`);
-      }
-    }
-    expect(bad, `dangling navigation targets:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('no user-facing object is stranded without a way to reach it', () => {
-    // Campaigns, Contracts, Products, Forecasts and Tasks each shipped with
-    // views, seed data and automation — and no entry point. The only way to
-    // see a task was to open the record it hung off.
-    const reachable = new Set(allNodes.map((n) => n.objectName).filter(Boolean));
-    const stranded = objects
-      .map((o) => o.name)
-      .filter((name: string) =>
-        // Line items and junctions are edited inside their parent, never
-        // reached on their own. `_attendee` joined the list with `crm_event`
-        // (#592): an attendee row is reached through the meeting it belongs to,
-        // and it is `controlled_by_parent`, so a nav entry would offer a list
-        // whose every row derives its visibility from a record you got to some
-        // other way.
-        !/_line_item$|_member$|_attendee$/.test(name) && !reachable.has(name));
-    expect(stranded, `objects with no navigation entry:\n  ${stranded.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every dashboard is reachable', () => {
-    const reachable = new Set(allNodes.map((n) => n.dashboardName).filter(Boolean));
-    const stranded = dashboards.map((d) => d.name).filter((n: string) => !reachable.has(n));
-    expect(stranded, `dashboards nobody can open:\n  ${stranded.join('\n  ')}`).toEqual([]);
-  });
-
-  it('widgets that window the dateRange field opt out of the injected date range', () => {
-    // The runtime injects the dashboard-level dateRange into every widget
-    // query (framework#2501) unless the widget opts out via `filterBindings`.
-    // A widget that carries its OWN window on the same field gets the two
-    // ANDed: the Executive "Total Revenue (YTD)" tile silently showed
-    // this-quarter revenue (781k instead of the year's 2.6M), and the
-    // "last 12 months" trend chart collapsed to the current quarter's points.
-    // Self-described windows ("YTD", "QTD", "last 12 months") must not follow
-    // the picker — their titles don't.
-    const bad: string[] = [];
-    for (const d of dashboards) {
-      const rangeField = d.dateRange?.field;
-      if (!rangeField) continue;
-      for (const w of d.widgets ?? []) {
-        if (w.filter?.[rangeField] === undefined) continue;
-        if (w.filterBindings?.dateRange !== false) {
-          bad.push(`${d.name}/${w.id}: filters on "${rangeField}" but still binds the dashboard dateRange`);
-        }
-      }
-    }
-    expect(bad, `widgets fighting the dashboard date picker:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every navigation node has a zh-CN label', () => {
-    // The groups were translated and the leaves were not, so the sidebar read
-    // half Chinese, half English.
-    const zh = packFor('zh-CN');
-    // Assert, don't skip: the old lookup used the wrong bundle shape, found no
-    // pack, and returned early — a green test that checked nothing.
-    expect(zh, 'no zh-CN locale pack found in stack.translations').toBeTruthy();
-    const bad: string[] = [];
-    for (const app of apps) {
-      const nav = zh?.apps?.[app.name]?.navigation ?? {};
-      for (const n of navNodes(app)) {
-        if (n.id && !nav[n.id]?.label) bad.push(`${app.name}/${n.id}`);
-      }
-    }
-    expect(bad, `navigation nodes with no zh-CN label:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-describe('dashboard actions land on real routes', () => {
-  /**
-   * A widget/header action with `actionType: 'url'` is pushed into the router
-   * verbatim (pushState, no normalisation). The 16.1.0 console serves in-app
-   * pages ONLY under `/apps/:appName/…` with these children (verified against
-   * the router in @objectstack/console):
-   *   :objectName | :objectName/view/:viewId | :objectName/record/:recordId |
-   *   dashboard/:dashboardName | report/:reportName | page/:pageName
-   * There is no top-level `/objects/…` or `/reports/…` route — 16 KPI tiles
-   * navigated to dead pages that way (issue #527).
-   */
-  const apps: AnyRec[] = (stack as any).apps ?? [];
-  const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
-  const reports: AnyRec[] = (stack as any).reports ?? [];
-  const appNames = new Set(apps.map((a) => a.name));
-  const reportNames = new Set(reports.map((r) => r.name));
-  const dashboardNames = new Set(dashboards.map((d) => d.name));
-  const pageNames = new Set(pages.map((p) => p.name));
-
-  /** List-view names per object (the default list plus every named listView). */
-  const viewNamesByObject = new Map<string, Set<string>>();
-  for (const v of views) {
-    const objectName = v.list?.data?.object;
-    if (!objectName) continue;
-    const known = viewNamesByObject.get(objectName) ?? new Set<string>();
-    if (v.list?.name) known.add(v.list.name);
-    for (const lv of Object.values(v.listViews ?? {}) as AnyRec[]) {
-      if (lv?.name) known.add(lv.name);
-    }
-    viewNamesByObject.set(objectName, known);
-  }
-
-  /** Returns undefined when the URL resolves, else why it doesn't. */
-  const routeError = (url: string): string | undefined => {
-    if (/^https?:\/\//.test(url) || url.startsWith('//')) return undefined; // external
-    const [path] = url.split(/[?#]/);
-    const seg = path.split('/').filter(Boolean);
-    if (seg[0] !== 'apps') return 'in-app URLs must start with /apps/<appName>/';
-    if (!appNames.has(seg[1])) return `app "${seg[1]}" is not defined`;
-    const [head, second, third] = seg.slice(2);
-    if (head === 'dashboard') {
-      return dashboardNames.has(second) ? undefined : `dashboard "${second}" is not defined`;
-    }
-    if (head === 'report') {
-      return reportNames.has(second) ? undefined : `report "${second}" is not defined`;
-    }
-    if (head === 'page') {
-      return pageNames.has(second) ? undefined : `page "${second}" is not defined`;
-    }
-    // `<objectName>` optionally followed by `view/<viewId>` (record deep links
-    // carry ids, not metadata — dashboards don't author those).
-    const objectOk = objectNames.has(head) || PLATFORM_OBJECTS.has(head);
-    if (!objectOk) return `object "${head}" is not defined`;
-    if (second === undefined) return undefined;
-    if (second !== 'view') return `unknown object sub-route "${second}"`;
-    return viewNamesByObject.get(head)?.has(third)
-      ? undefined
-      : `object "${head}" has no list view "${third}"`;
-  };
-
-  it('every url-type widget and header action resolves', () => {
-    const bad: string[] = [];
-    for (const d of dashboards) {
-      const actions = [
-        ...(d.header?.actions ?? []),
-        ...(d.widgets ?? []),
-      ] as AnyRec[];
-      for (const a of actions) {
-        if (!a.actionUrl || (a.actionType ?? 'url') !== 'url') continue;
-        const err = routeError(a.actionUrl);
-        if (err) bad.push(`${d.name}/${a.id ?? a.label}: ${a.actionUrl} — ${err}`);
-      }
-    }
-    expect(bad, `dashboard actions navigating to dead routes:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  /**
-   * The rule above currently iterates nothing: no dashboard declares a
-   * widget-level or header action any more, because the console renders no
-   * drill-through for a dataset-bound widget (measured on 16.1.0 — a KPI tile
-   * emits no link, no button and no icon, its cursor stays `auto`, and clicking
-   * it does not navigate). Removing the config was the point of this PR.
-   *
-   * A guard that iterates an empty list is a guard that passes without checking
-   * anything — the exact failure the navigation rule above was written to fix.
-   * So the checker is exercised directly here: it must still reject the two
-   * shapes #527 found in the wild, and still accept a real one. If someone
-   * re-adds an action later, the rule above resumes covering it for real.
-   */
-  it('the route checker still rejects the shapes #527 found', () => {
-    expect(routeError('/objects/opportunity?filter=open')).toBeTruthy();
-    expect(routeError('/reports/revenue')).toBeTruthy();
-    expect(routeError('/apps/no_such_app/crm_lead')).toBeTruthy();
-    expect(routeError('/apps/crm_enterprise/crm_lead/view/no_such_view')).toBeTruthy();
-    expect(routeError('/apps/crm_enterprise/crm_lead')).toBeUndefined();
-    expect(routeError('https://example.com/anything')).toBeUndefined();
-  });
-});
-
-describe('filter template tokens are resolvable', () => {
-  /**
-   * Token support differs per data path — and the difference MOVED under us at
-   * the 17.0 upgrade, so the shape of this rule changed with it (#730).
-   *
-   * - LIST-VIEW data path (`/api/v1/data/...`): resolves user tokens AND date
-   *   macros. The date-macro half is new. Through 16.1.0 nothing on the server
-   *   substituted placeholders on the read path, so a macro shipped as a
-   *   literal string — and because `'2026-…' < '{…'` is lexicographically
-   *   TRUE, a `< {180_days_ago}` filter matched freshly-reviewed articles
-   *   (inverted semantics, not merely empty). That is why this rule used to
-   *   read "user tokens ONLY", and why views were written as operator-only
-   *   filters plus a sort.
-   *
-   *   `resolveFilterTokens()` was wired into the ObjectQL READ path in
-   *   17.0.0-rc.0 (objectql `feat(filters): evaluate {filter-token}
-   *   placeholders server-side`, #3582) — ahead of the middleware chain,
-   *   covering `find`/`findOne`/`count`/`aggregate`, which is the one gate
-   *   every server-side read passes through, saved-view filters included. The
-   *   lexicographic hazard above is gone in BOTH directions now: a known token
-   *   is substituted before the driver sees it, and an unknown one throws
-   *   `Unresolvable filter placeholder` naming the near-miss fix instead of
-   *   comparing as text.
-   *
-   *   Measured on the pinned 17.0.0-rc.2, not inferred from the changelog:
-   *   `test/forecast-current-quarter-view.test.ts` runs a shipped view filter
-   *   through a real engine and pins one quarter selected out of three, plus
-   *   the throw on the retired `{this_quarter_start}` spelling.
-   *
-   * - ANALYTICS path (dashboard widgets / dataset reports,
-   *   `/api/v1/analytics/...`): resolves the DATE_MACRO_TOKENS vocabulary (the
-   *   YTD revenue widget returns a non-zero sum, impossible with a literal
-   *   token), but still NO user token — `{current_user}` and even
-   *   `{current_user_id}` match no owner (see crm.app.ts's My Work note). That
-   *   half is unchanged and tracked at #510, so the two rules below remain
-   *   asymmetric — just not in the direction they were written for.
-   */
-  const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
-  const reports: AnyRec[] = (stack as any).reports ?? [];
-
-  const USER_TOKENS = new Set(['current_user_id', 'current_org_id']);
-
-  /** Yield every string inside an arbitrarily nested filter value. */
-  function* filterStrings(node: unknown): Generator<string> {
-    if (typeof node === 'string') { yield node; return; }
-    if (Array.isArray(node)) { for (const item of node) yield* filterStrings(item); return; }
-    if (node && typeof node === 'object') {
-      for (const value of Object.values(node)) yield* filterStrings(value);
-    }
-  }
-
-  const badTokensIn = (
-    where: string,
-    filter: unknown,
-    allowed: (token: string) => boolean,
-    bad: string[],
-  ) => {
-    for (const s of filterStrings(filter)) {
-      for (const m of s.matchAll(/\{([^{}]+)\}/g)) {
-        if (!allowed(m[1])) bad.push(`${where}: unresolvable token "{${m[1]}}"`);
-      }
-    }
-  };
-
-  it('list view and page filters only use user tokens or date macros', () => {
-    const allowed = (t: string) => USER_TOKENS.has(t) || isDateMacroToken(t);
-    const bad: string[] = [];
-    for (const v of views) {
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) badTokensIn(`view "${list.name ?? 'default'}"`, list.filter, allowed, bad);
-    }
-    for (const p of pages) {
-      badTokensIn(`page "${p.name}"`, p.interfaceConfig?.filterBy, allowed, bad);
-      // Page COMPONENTS carry filters too, and did not used to be walked. Sales
-      // Home now embeds `list-view` blocks whose filters come off the saved
-      // views (#771) — those are checked above as views, but a hand-written
-      // component filter would have reached the same data path unexamined.
-      for (const c of [...walk(p.regions), ...walk(p.slots)]) {
-        badTokensIn(`page "${p.name}"/${c.id ?? c.type}`, c.properties?.filter, allowed, bad);
-        badTokensIn(`page "${p.name}"/${c.id ?? c.type}`, c.dataSource?.filter, allowed, bad);
-      }
-    }
-    expect(bad, `unresolvable view/page filter tokens:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('the widened list-view rule still rejects what neither path resolves', () => {
-    // Widening a rule is how a rule dies. `{this_quarter_start}` is the exact
-    // spelling #515 removed from `this_quarter_forecasts`, and the engine now
-    // throws on it — so it must stay rejected HERE too, at author time, rather
-    // than being waved through by the new date-macro allowance. `{current_user}`
-    // covers the other near-miss (the resolvable spelling is `{current_user_id}`).
-    const allowed = (t: string) => USER_TOKENS.has(t) || isDateMacroToken(t);
-    const bad: string[] = [];
-    badTokensIn('probe', [
-      { value: '{this_quarter_start}' },
-      { value: '{current_user}' },
-      { value: '{not_a_token}' },
-    ], allowed, bad);
-    expect(bad).toHaveLength(3);
-
-    // …and still accepts both classes it is now supposed to accept.
-    const good: string[] = [];
-    badTokensIn('probe', [
-      { value: '{current_user_id}' },
-      { value: '{current_quarter_start}' },
-      { value: '{30_days_ago}' },
-    ], allowed, good);
-    expect(good).toEqual([]);
-  });
-
-  it('dashboard widget and report filters only use date macros', () => {
-    const allowed = (t: string) => isDateMacroToken(t);
-    const bad: string[] = [];
-    for (const d of dashboards) {
-      for (const w of d.widgets ?? []) badTokensIn(`${d.name}/${w.id}`, w.filter, allowed, bad);
-    }
-    for (const r of reports) badTokensIn(`report "${r.name}"`, r.runtimeFilter, allowed, bad);
-    expect(bad, `unresolvable analytics filter tokens:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
 describe('object references outside views resolve', () => {
   const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
   const actions: AnyRec[] = (stack as any).actions ?? [];
@@ -861,393 +422,6 @@ describe('object references outside views resolve', () => {
       }
     }
     expect(bad, `dangling action param objects:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-describe('list-level action references resolve', () => {
-  const actions: AnyRec[] = (stack as any).actions ?? [];
-  const actionNames = new Set(actions.map((a) => a.name));
-
-  /**
-   * There is NO builtin escape hatch (@objectstack 17).
-   *
-   * This guard used to whitelist `edit` / `delete` / `view` as "affordances the
-   * list renderer provides without an Action def". 17.0's `action-name-undefined`
-   * validator rule refuses exactly that: a `rowActions` string naming no defined
-   * action is reported as a dead affordance — "the button renders and does
-   * nothing when clicked" — and `os validate` fails the build.
-   *
-   * So the whitelist was a trap: it let metadata pass CI that the platform then
-   * rejected. Row-menu entries come from an Action declaring
-   * `locations: ['list_item']`, which auto-injects them; naming one here as a
-   * string is the legacy path the next test forbids. Between the two rules,
-   * `rowActions` has no correct use in this app today — which is why no view
-   * declares it.
-   */
-  it('every rowAction / bulkAction names a defined action', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        for (const name of [...(list.rowActions ?? []), ...(list.bulkActions ?? [])]) {
-          if (typeof name === 'string' && !actionNames.has(name)) {
-            bad.push(`view "${list.name ?? 'default'}": action "${name}" is not defined`);
-          }
-        }
-      }
-    }
-    expect(bad, `dangling list action references:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('no rowAction repeats an action that already declares list_item placement', () => {
-    // An Action with `locations: ['list_item']` auto-injects its row-menu
-    // entry. Naming it AGAIN as a rowActions string goes through objectui's
-    // legacy path, which dispatches the string as an action TYPE — producing
-    // a second, dead menu item (issue #535 / objectstack-ai/objectui#2960).
-    const listItemActions = new Set(
-      actions.filter((a) => (a.locations ?? []).includes('list_item')).map((a) => a.name),
-    );
-    const bad: string[] = [];
-    for (const v of views) {
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        for (const name of list.rowActions ?? []) {
-          if (typeof name === 'string' && listItemActions.has(name)) {
-            bad.push(`view "${list.name ?? 'default'}": "${name}" duplicates its list_item auto-injection`);
-          }
-        }
-      }
-    }
-    expect(bad, `redundant string rowActions:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  /**
-   * Bulk dispatch (`bulkActionDefs` / `bulkActions`) is pinned in
-   * `test/bulk-action-dispatch.test.ts`.
-   *
-   * A guard forbidding `mass_update_stage` on any list used to live here: the
-   * selection-bar button was found to be a client-side no-op (#588), so the
-   * wiring was treated as the defect. #508 disproved that premise — the console
-   * delivers a selection through `bulkActionDefs` + `execution: 'aggregate'`
-   * (objectstack-ai/objectstack#5568), and the button was inert because the
-   * DECLARATION had been removed. The guard is inverted rather than deleted,
-   * and moved: this file is one edit from the 100KB hygiene ceiling.
-   */
-});
-
-describe('row colors and kanban groups key off real option values', () => {
-  const viewObjectOf = (v: AnyRec): string | undefined =>
-    v.list?.data?.object ?? v.form?.data?.object ?? v.object;
-
-  it('rowColor keys on select fields are actual option values', () => {
-    // Task rows were colored by `critical`/`medium` — Case values. Task
-    // priority is low/normal/high/urgent, so urgent rows rendered uncolored.
-    const bad: string[] = [];
-    for (const v of views) {
-      const objectName = viewObjectOf(v);
-      const objDef = objects.find((o) => o.name === objectName);
-      if (!objDef) continue;
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        const rc = list.rowColor;
-        if (!rc?.field) continue;
-        const options = objDef.fields?.[rc.field]?.options;
-        if (!Array.isArray(options) || !options.length) continue; // not a select — booleans/numbers are fine
-        const values = new Set(options.map((o: AnyRec) => String(o.value)));
-        for (const key of Object.keys(rc.colors ?? {})) {
-          if (!values.has(key)) {
-            bad.push(`view "${list.name ?? 'default'}": rowColor key "${key}" is not an option of ${objectName}.${rc.field}`);
-          }
-        }
-      }
-    }
-    expect(bad, `dangling rowColor keys:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('kanban groupByField is a select field with options', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const objectName = viewObjectOf(v);
-      const objDef = objects.find((o) => o.name === objectName);
-      if (!objDef) continue;
-      const lists = [v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[];
-      for (const list of lists) {
-        const groupBy = list.kanban?.groupByField;
-        if (!groupBy) continue;
-        const field = objDef.fields?.[groupBy];
-        if (!field) bad.push(`view "${list.name}": kanban groups by missing field "${groupBy}"`);
-        else if (!Array.isArray(field.options) || !field.options.length) {
-          bad.push(`view "${list.name}": kanban groupByField "${groupBy}" has no options`);
-        }
-      }
-    }
-    expect(bad, `broken kanban groupings:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-/**
- * Stage enumerations are COMPLETE, not merely valid.
- *
- * The two guards above answer "is every value written here a real option?" —
- * a subset check. Nothing asked the converse, "is every real option written
- * here?", and that is the half that broke (#759): `needs_analysis` is one of
- * the seven canonical `OPPORTUNITY_STAGE_OPTIONS` and was absent from BOTH the
- * detail page's stage path and the Open Deals `rowColor` map. Six of seven
- * stages worked, so both surfaces looked fine on every deal that happened not
- * to be in Needs Analysis — the deal that WAS lit up no step on the path and
- * got no row tint, which reads to a user as corrupted data rather than as
- * missing metadata.
- *
- * The expectation is derived from `OPPORTUNITY_STAGE_OPTIONS` itself — the
- * single source `crm_opportunity.stage` and the `mass_update_stage` action are
- * both built from — so an EIGHTH stage cannot ship half-covered. A hand-copied
- * list here would need the same edit as the metadata it guards, and would
- * therefore be forgotten in the same commit.
- */
-describe('every canonical opportunity stage reaches the UI that enumerates stages', () => {
-  const canonical = OPPORTUNITY_STAGE_OPTIONS.map((o) => String(o.value));
-
-  /** Every `record:path` bound to `crm_opportunity.stage`, as [pageName, values]. */
-  const stagePaths = (): [string, string[]][] =>
-    pages
-      .filter((page) => page.object === 'crm_opportunity')
-      .flatMap((page) =>
-        [...walk(page.regions), ...walk(page.slots)]
-          .filter((c) => c.type === 'record:path' && c.properties?.statusField === 'stage')
-          .map((c): [string, string[]] => [
-            `${page.name} / ${c.id}`,
-            (c.properties?.stages ?? []).map((s: AnyRec) => String(s.value)),
-          ]),
-      );
-
-  /** Every list `rowColor` keyed on `crm_opportunity.stage`, as [viewName, colors]. */
-  const stageRowColors = (): [string, AnyRec][] =>
-    views
-      .filter((v) => (v.list?.data?.object ?? v.object) === 'crm_opportunity')
-      .flatMap((v) =>
-        ([v.list, ...Object.values(v.listViews ?? {})].filter(Boolean) as AnyRec[])
-          .filter((list) => list.rowColor?.field === 'stage')
-          .map((list): [string, AnyRec] => [
-            `view "${list.name ?? 'default'}"`,
-            list.rowColor?.colors ?? {},
-          ]),
-      );
-
-  it('the canonical stage list is the set the object validates against', () => {
-    // Guards the guard twice over: an empty or renamed constant would make
-    // every assertion below pass by comparing against nothing, and a `stage`
-    // field that stopped being built from this constant would leave the two
-    // sites below chasing a list the runtime no longer enforces.
-    expect(canonical.length, 'OPPORTUNITY_STAGE_OPTIONS is empty').toBeGreaterThan(0);
-    const objDef = objects.find((o) => o.name === 'crm_opportunity');
-    const fieldValues = (objDef?.fields?.stage?.options ?? []).map((o: AnyRec) => String(o.value));
-    expect([...fieldValues].sort(), 'crm_opportunity.stage no longer mirrors OPPORTUNITY_STAGE_OPTIONS')
-      .toEqual([...canonical].sort());
-  });
-
-  it('every stage has a step on the opportunity stage path', () => {
-    const sites = stagePaths();
-    // A renamed component type or statusField would silently empty this list
-    // and turn the assertion below into a no-op — exactly how the navigation
-    // guard in this file spent its life passing.
-    expect(sites.length, 'no record:path bound to crm_opportunity.stage was found').toBeGreaterThan(0);
-
-    const bad: string[] = [];
-    for (const [site, values] of sites) {
-      for (const stage of canonical) {
-        if (!values.includes(stage)) bad.push(`${site}: no path step for stage "${stage}"`);
-      }
-      for (const value of values) {
-        if (!canonical.includes(value)) bad.push(`${site}: path step "${value}" is not a canonical stage`);
-      }
-    }
-    expect(bad, `incomplete stage paths:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every stage has a row colour in the opportunity list views that tint by stage', () => {
-    const sites = stageRowColors();
-    expect(sites.length, 'no rowColor keyed on crm_opportunity.stage was found').toBeGreaterThan(0);
-
-    const bad: string[] = [];
-    for (const [site, colors] of sites) {
-      const keys = Object.keys(colors);
-      for (const stage of canonical) {
-        if (!keys.includes(stage)) bad.push(`${site}: no rowColor entry for stage "${stage}"`);
-      }
-      for (const key of keys) {
-        if (!canonical.includes(key)) bad.push(`${site}: rowColor key "${key}" is not a canonical stage`);
-      }
-    }
-    expect(bad, `incomplete stage row colours:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('stage colours stay distinguishable from one another', () => {
-    // The point of the map is that a rep can tell two stages apart at a
-    // glance; two stages sharing a hex make the tint say nothing. (#759 was
-    // nearly fixed by copying `needs_analysis`'s `#FFD700` out of
-    // _picklists.ts, one hue step from proposal's `#f59e0b`.)
-    const bad: string[] = [];
-    for (const [site, colors] of stageRowColors()) {
-      const seen = new Map<string, string>();
-      for (const key of Object.keys(colors)) {
-        const hex = String(colors[key]).toLowerCase();
-        const owner = seen.get(hex);
-        if (owner) bad.push(`${site}: "${key}" and "${owner}" are both ${hex}`);
-        else seen.set(hex, key);
-      }
-    }
-    expect(bad, `stages sharing a row colour:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-describe('picklist values never reach the UI unresolved', () => {
-  /**
-   * A select field stores a VALUE (`ms`, `waiting_customer`, `existing_upgrade`)
-   * and the UI resolves it to the locale's label at render time. Every #461
-   * defect was that resolution failing, in one of two ways — a lookup that
-   * misses (the translation is keyed by something that is not an option value)
-   * or a lookup that never happens (a formula splices the stored value straight
-   * into a string). Both are invisible in review and only surface as English —
-   * or worse, `closed_won` — sitting in an otherwise-translated screen.
-   */
-  const selectOptionsOf = (objDef: AnyRec, field: string): string[] | undefined => {
-    const options = objDef.fields?.[field]?.options;
-    return Array.isArray(options) && options.length
-      ? options.map((o: AnyRec) => String(o.value))
-      : undefined;
-  };
-
-  it('option translations are keyed by option VALUE, not by English label', () => {
-    // Opportunity `type` was keyed by label ('Existing Customer - Upgrade':
-    // '老客户升级'). The resolver matches by value, so deal type rendered in
-    // English on every non-en locale while `stage`/`status` — keyed by value —
-    // translated fine, which is exactly what made it hard to spot.
-    const bad: string[] = [];
-    for (const [locale, pack] of localePacks) {
-      for (const [objName, objT] of Object.entries<AnyRec>(pack?.objects ?? {})) {
-        const objDef = objects.find((o) => o.name === objName);
-        if (!objDef) continue; // covered by the key-resolution test below
-        for (const [fieldName, fieldT] of Object.entries<AnyRec>(objT?.fields ?? {})) {
-          if (!fieldT?.options) continue;
-          const values = selectOptionsOf(objDef, fieldName);
-          if (!values) {
-            bad.push(`${locale}: ${objName}.${fieldName} translates options but the field has none`);
-            continue;
-          }
-          for (const key of Object.keys(fieldT.options)) {
-            if (!values.includes(key)) {
-              bad.push(`${locale}: ${objName}.${fieldName} option key "${key}" is not an option value`);
-            }
-          }
-        }
-      }
-    }
-    expect(bad, `option translations that resolve to nothing:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('translated object and field keys name real objects and fields', () => {
-    // A translation keyed to a renamed/removed field is dead weight that reads
-    // as coverage — the bundle looks complete while the screen stays English.
-    const bad: string[] = [];
-    for (const [locale, pack] of localePacks) {
-      for (const [objName, objT] of Object.entries<AnyRec>(pack?.objects ?? {})) {
-        if (!objectNames.has(objName)) {
-          bad.push(`${locale}: translates "${objName}", which is not a defined object`);
-          continue;
-        }
-        const known = fieldsOf(objName);
-        for (const fieldName of Object.keys(objT?.fields ?? {})) {
-          if (!known.includes(fieldName)) {
-            bad.push(`${locale}: "${objName}" has no field "${fieldName}"`);
-          }
-        }
-      }
-    }
-    expect(bad, `translations keyed to nothing:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('no formula renders a select field into its output string', () => {
-    // `full_name` = joinNonEmpty([salutation, first_name, last_name]) printed
-    // "ms Emily Davis"; opportunity `display_title` = name + " - " + stage
-    // titled deals "Enterprise Deal - closed_won". The formula language has no
-    // option-label lookup (only upper/lower), so a select simply cannot be
-    // rendered from a formula — it has to stay its own field.
-    //
-    // Only RENDERING is flagged. Branching on a select (`record.stage ==
-    // "closed_won" ? …`) never puts the value on screen and stays legal.
-    const rendersSelect = (source: string, field: string): boolean => {
-      const ref = `record\\.${field}\\b`;
-      // joinNonEmpty([...]) — every element lands in the output string.
-      for (const m of source.matchAll(/joinNonEmpty\(\s*\[([^\]]*)\]/g)) {
-        if (new RegExp(ref).test(m[1])) return true;
-      }
-      // String concatenation on either side: `" - " + record.x` / `record.x + " - "`.
-      return new RegExp(`(["']\\s*\\+\\s*${ref})|(${ref}\\s*\\+\\s*["'])`).test(source);
-    };
-
-    const bad: string[] = [];
-    for (const objDef of objects) {
-      for (const [fieldName, field] of Object.entries<AnyRec>(objDef.fields ?? {})) {
-        // `expression` is `{ dialect, source }` once the F tag is evaluated.
-        const source: unknown = field?.expression?.source ?? field?.expression;
-        if (typeof source !== 'string') continue;
-        for (const other of Object.keys(objDef.fields ?? {})) {
-          if (!selectOptionsOf(objDef, other)) continue;
-          if (rendersSelect(source, other)) {
-            bad.push(`${objDef.name}.${fieldName} renders select "${other}": ${source}`);
-          }
-        }
-      }
-    }
-    expect(bad, `formulas printing raw select values:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-describe('every named list view is reachable', () => {
-  const apps: AnyRec[] = (stack as any).apps ?? [];
-  const navViewNames = new Set(
-    apps.flatMap((app) =>
-      (app.navigation ?? []).flatMap(function walk(n: AnyRec): string[] {
-        return [...(n.viewName ? [n.viewName] : []), ...(n.children ?? []).flatMap(walk)];
-      })),
-  );
-
-  it('every listViews entry is referenced by the switcher tabs or app navigation', () => {
-    // With no `tabs` declared, the data-mode switcher lists every saved view
-    // automatically (ADR-0047) — nothing to check. But once a view curates a
-    // `tabs` array, only the listed views render: seven working queues
-    // (renewals_due, at_risk_accounts, stale_opportunities,
-    // closing_this_quarter, sla_at_risk, todays_tasks, overdue_tasks) shipped
-    // outside their list's tabs — defined, tested, unreachable.
-    const bad: string[] = [];
-    for (const v of views) {
-      const tabs = v.list?.tabs;
-      if (!Array.isArray(tabs) || !tabs.length) continue;
-      const tabViews = new Set(tabs.map((t: AnyRec) => t.view).filter(Boolean));
-      for (const [key, def] of Object.entries(v.listViews ?? {}) as [string, AnyRec][]) {
-        const name = def.name ?? key;
-        if (!tabViews.has(name) && !navViewNames.has(name)) {
-          bad.push(`list view "${name}" (${def.data?.object}) is excluded from its list's tabs and has no navigation entry`);
-        }
-      }
-    }
-    expect(bad, `unreachable list views:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every switcher tab points at a defined view', () => {
-    const bad: string[] = [];
-    for (const v of views) {
-      const defined = new Set([
-        v.list?.name,
-        ...Object.entries(v.listViews ?? {}).map(([key, def]) => (def as AnyRec)?.name ?? key),
-      ].filter(Boolean));
-      for (const t of v.list?.tabs ?? []) {
-        if (t.view && !defined.has(t.view)) {
-          bad.push(`tab "${t.name}" targets undefined view "${t.view}"`);
-        }
-      }
-    }
-    expect(bad, `dangling tabs:\n  ${bad.join('\n  ')}`).toEqual([]);
   });
 });
 
@@ -1388,236 +562,6 @@ describe('page templates and record components stay inside their record context'
   });
 });
 
-describe('dashboard date ranges window a field the query layer can actually compare', () => {
-  /**
-   * A dashboard `dateRange` is ANDed into EVERY widget query, so if the
-   * comparison cannot match, the whole dashboard renders zeros. That is #460:
-   * the Service dashboard windowed `crm_case.created_date` — a
-   * `Field.datetime()` — and opened with every KPI at 0 and every chart empty,
-   * with 38 cases in the system.
-   *
-   * The mechanism is a storage disagreement, not a bad preset. On SQLite,
-   * `driver-sql` 16.1.0 coerces datetime filter values to epoch-millisecond
-   * INTEGERs (`coerceFilterValue`), documenting the assumption that datetime
-   * columns hold INTEGER ms. In this app they hold ISO TEXT — including the
-   * platform's own `created_at`/`updated_at`. SQLite orders every INTEGER
-   * before every TEXT, so `datetime_col >= <int>` is true for all rows and
-   * `datetime_col <= <int>` is true for none. Measured against the running
-   * 16.1.0 console: `$gte` alone → all 38 cases, `$lte` alone → 0, both → 0,
-   * in every date format tried.
-   *
-   * `Field.date()` is unaffected (TEXT `YYYY-MM-DD` on both sides), which is
-   * why the CRM/Sales/Executive dashboards — all windowing `close_date` — show
-   * data. So the rule is about the FIELD TYPE, not the preset: a dashboard may
-   * only window a `date` field. Note this deliberately fails if someone
-   * restores the Service dashboard's `dateRange` before the driver is fixed.
-   */
-  const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
-  const datasets: AnyRec[] = (stack as any).datasets ?? [];
-
-  /** Objects the dashboard's widgets aggregate over, via their datasets. */
-  const objectsBehind = (d: AnyRec): string[] => {
-    const names = new Set<string>();
-    for (const w of d.widgets ?? []) {
-      const ds = datasets.find((x) => x.name === w.dataset);
-      if (ds?.object) names.add(ds.object);
-    }
-    return [...names];
-  };
-
-  const fieldType = (objectName: string, field: string): string | undefined =>
-    objects.find((o) => o.name === objectName)?.fields?.[field]?.type;
-
-  it('no dashboard windows a datetime field', () => {
-    const bad: string[] = [];
-    for (const d of dashboards) {
-      const field = d.dateRange?.field;
-      if (!field) continue;
-      for (const obj of objectsBehind(d)) {
-        const type = fieldType(obj, field);
-        if (type === 'datetime') {
-          bad.push(
-            `${d.name}: dateRange windows ${obj}.${field}, a datetime field — ` +
-            `the $lte bound matches no rows, so every widget renders empty`,
-          );
-        }
-      }
-    }
-    expect(bad, `dashboards that will render empty:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every dashboard dateRange field exists on the objects its widgets aggregate', () => {
-    // A range field that no underlying object defines is silently dropped or
-    // errors per widget — either way the picker cannot do what it claims.
-    const bad: string[] = [];
-    for (const d of dashboards) {
-      const field = d.dateRange?.field;
-      if (!field) continue;
-      const behind = objectsBehind(d);
-      if (!behind.length) continue;
-      const resolves = behind.filter((obj) => fieldType(obj, field) !== undefined);
-      if (!resolves.length) {
-        bad.push(`${d.name}: dateRange field "${field}" exists on none of ${behind.join(', ')}`);
-      }
-    }
-    expect(bad, `dangling dashboard date-range fields:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-/**
- * Action labels are the most visible strings in the app — they render in row
- * menus, on detail-page headers and in list toolbars. They also had no guard,
- * so #494's i18n audit missed the whole class: three actions
- * (`enroll_leads`, `schedule_followup`, `generate_quote`) shipped with no label
- * entry in ANY locale pack, and `log_meeting` had one only in zh-CN.
- *
- * The gap only became visible as recent PRs surfaced those actions:
- * `generate_quote` moved onto the opportunity header (#515), `enroll_leads`
- * was introduced with the campaign-enrollment screen flow (#536), and
- * `schedule_followup` became the row menu's sole remaining entry once the dead
- * duplicate was removed (#537). A missing entry is not an error — it silently
- * falls back to the English `label` in code — which is exactly why it needs a
- * test rather than a warning.
- */
-describe('action labels are translated in every locale', () => {
-  it('every action has a label entry in every locale pack', () => {
-    expect(localePacks.length, 'no locale packs found in stack.translations').toBeGreaterThan(0);
-    const bad: string[] = [];
-    for (const action of stackActions) {
-      const { name, objectName } = action;
-      if (!name || !objectName) continue;
-      for (const [locale, pack] of localePacks) {
-        const label = pack?.objects?.[objectName]?._actions?.[name]?.label;
-        if (!label) bad.push(`${locale}: ${objectName}._actions.${name}.label`);
-      }
-    }
-    expect(bad, `actions with no translated label:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('confirmText and successMessage are translated wherever the action declares them', () => {
-    // A half-translated action is worse than an untranslated one: the menu item
-    // reads Chinese and then the confirm dialog answers in English.
-    const bad: string[] = [];
-    for (const action of stackActions) {
-      const { name, objectName } = action;
-      if (!name || !objectName) continue;
-      for (const key of ['confirmText', 'successMessage'] as const) {
-        if (!action[key]) continue;
-        for (const [locale, pack] of localePacks) {
-          const entry = pack?.objects?.[objectName]?._actions?.[name];
-          if (entry && !entry[key]) bad.push(`${locale}: ${objectName}._actions.${name}.${key}`);
-        }
-      }
-    }
-    expect(bad, `action strings declared in code but not translated:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-});
-
-/**
- * Select fields — the field's own label AND every option label — are translated
- * in every locale.
- *
- * The sibling guard above does this for action labels (#494). It was never
- * extended to select fields, and `crm_lead.disqualification_reason` (#631) is
- * what that costs: a `required` field on eight lead forms with no entry in ANY
- * of the four bundles, so the form rendered `not_a_fit` / `no_budget` /
- * `wrong_persona` as raw stored VALUES in the middle of an otherwise fully
- * translated screen.
- *
- * It stayed invisible for the same reason #494's did: a missing entry is not an
- * error. The resolver falls back to the English `label` in code, and in `en`
- * that fallback is indistinguishable from a correct translation — so the only
- * locale a reviewer is likely to open is the one locale where the bug cannot be
- * seen. That is a class a test catches and an eyeball does not.
- *
- * The two neighbouring guards in `picklist values never reach the UI unresolved`
- * check that the entries which DO exist point at something real. This one checks
- * the other direction: that they exist at all. Both are needed — a bundle can be
- * perfectly well-formed and still be empty.
- */
-describe('select fields are translated in every locale', () => {
-  /**
-   * There is no exemption list any more, and re-introducing one is a regression.
-   *
-   * This guard shipped with a `PENDING_SELECT_LABELS` ledger: 34 fields over 12
-   * objects, 111 (locale, field) pairs, ~380 option labels of pre-existing debt
-   * that #631 could not fix without burying the one field its PR was about. The
-   * ledger was allowed to shrink and never grow, and #645 emptied it. It is gone
-   * rather than left behind as an empty object, along with the two assertions
-   * that existed only to keep it honest (stale rows, ghost rows) — an empty
-   * ledger is an invitation to add a row, and the assertions below say what the
-   * ledger was always converging on: every select field, every option value,
-   * every locale, no exceptions.
-   *
-   * A field added or extended from here on therefore fails on the PR that
-   * introduces it, which is the whole point.
-   */
-
-  /** Every authored select field, with its option VALUES (what the DB stores). */
-  const selectFields = objects.flatMap((obj) =>
-    Object.entries<AnyRec>(obj.fields ?? {})
-      .filter(([, f]) => Array.isArray(f?.options) && f.options.length)
-      .map(([fieldName, f]) => ({
-        key: `${obj.name}.${fieldName}`,
-        objectName: obj.name as string,
-        fieldName,
-        values: (f.options as AnyRec[]).map((o) => String(o.value)),
-      })),
-  );
-
-  /** The translation entry for one select field in one locale pack. */
-  const entryFor = (pack: AnyRec, objectName: string, fieldName: string): AnyRec | undefined =>
-    pack?.objects?.[objectName]?.fields?.[fieldName];
-
-  it('sees a non-trivial set of select fields and locales', () => {
-    // Guards the guard: `stack.objects` returning [] (or the translations
-    // bundle failing to flatten) would make every assertion below pass by
-    // checking nothing — which is exactly how the navigation guard in this file
-    // spent its life green.
-    expect(selectFields.length, 'no select fields discovered').toBeGreaterThanOrEqual(40);
-    expect(localePacks.length, 'no locale packs found in stack.translations').toBeGreaterThan(0);
-  });
-
-  it('every select field has a translated label in every locale', () => {
-    const bad: string[] = [];
-    for (const { key, objectName, fieldName } of selectFields) {
-      for (const [locale, pack] of localePacks) {
-        if (!entryFor(pack, objectName, fieldName)?.label) {
-          bad.push(`${locale}: ${objectName}.fields.${fieldName}.label`);
-        }
-      }
-    }
-    expect(
-      bad,
-      `select fields with no translated label:\n  ${bad.join('\n  ')}\n` +
-        'Add the entry to src/translations/<locale>.ts — a missing one silently ' +
-        'falls back to the English label in code.',
-    ).toEqual([]);
-  });
-
-  it('every option value has a translated label in every locale', () => {
-    // The half of the defect users actually notice. A missing field label reads
-    // as an odd column heading; a missing OPTION label puts the raw stored value
-    // (`not_a_fit`, `waiting_customer`) into a picklist a rep has to choose from.
-    const bad: string[] = [];
-    for (const { key, objectName, fieldName, values } of selectFields) {
-      for (const [locale, pack] of localePacks) {
-        const options = entryFor(pack, objectName, fieldName)?.options ?? {};
-        const missing = values.filter((v) => !options[v]);
-        if (missing.length) {
-          bad.push(`${locale}: ${objectName}.${fieldName} — ${missing.join(', ')}`);
-        }
-      }
-    }
-    expect(
-      bad,
-      `option values with no translated label:\n  ${bad.join('\n  ')}\n` +
-        'These render as the raw stored value in the picklist.',
-    ).toEqual([]);
-  });
-
-});
-
 /**
  * `form.data` is a data provider the form renderer never reads: a form binds to
  * its object and record through the route context, so the block only *looked*
@@ -1726,381 +670,5 @@ describe('app AI bindings resolve to a platform agent', () => {
       'app-authored agents were retired in #512 — author skills instead (ADR-0063 §2)',
     ).toEqual([]);
     expect(((stack as any).skills ?? []).length, 'no skills registered').toBeGreaterThan(0);
-  });
-});
-
-/**
- * EVERY locale is complete on every translatable surface, not just on select
- * options.
- *
- * The guard above covers ONE surface — picklist option labels. It says nothing
- * about the other four an authored bundle owns (field labels/help, page header
- * copy, dashboard widget titles, list-view empty states), and those are exactly
- * where the gaps sat: every page in the app rendered its header, nav label and
- * breadcrumb in English in three of four locales, and the Sales-dashboard
- * win/loss widgets did the same.
- *
- * Nothing in CI could have caught that. `pnpm lint` runs with `--skip-i18n`,
- * and `objectstack lint` exits 0 on warnings regardless — so the i18n rules
- * that DO find these gaps are switched off in the one place that would fail a
- * PR. This suite is the only gate, which is why the assertions live here rather
- * than in a lint config.
- *
- * This started as a zh-CN-only guard, because zh-CN was completed first ahead
- * of a customer trial while `en` / `ja-JP` / `es-ES` still carried the debt in
- * #645 and #494. That debt is now paid, so the guard covers all four — which
- * was always the stated finish line, and is the only version of it that stops
- * the next locale from drifting.
- *
- * Derived from `localePacks` rather than a hard-coded list: a fifth locale
- * added to the bundle is held to the same bar the day it appears, instead of
- * silently enjoying an exemption nobody wrote down.
- *
- * ### The surface list is derived from the METADATA, never from lint's categories
- *
- * The `_sections` assertion below was missing from the first version of this
- * guard, and its absence is the lesson the rest of the file should be read
- * against. The surfaces were chosen by reading off `objectstack lint`'s warning
- * categories — missing-option / -field / -navigation / -page / -action / -view
- * / -widget / -object. There is no `_sections` category, so sections were never
- * considered, and the guard inherited the linter's blind spot while reporting
- * that every locale was complete.
- *
- * What that cost: 83 of 85 section headings were untranslated in `ja-JP` and
- * `es-ES`, so a Japanese user read "Basic Information" and "Ownership & Status"
- * on every record page under an otherwise fully Japanese UI — while lint
- * reported zero i18n warnings for that locale (#683).
- *
- * A tool's warning list is a record of what someone thought to check, not of
- * what exists. Every assertion here therefore walks the metadata itself and
- * asks what it declares; none of them asks a linter what it noticed.
- */
-describe('every locale is complete on every authored surface', () => {
-  const packs = (): [string, AnyRec][] => {
-    // Guards the guard: an empty list would make every assertion below pass by
-    // checking nothing — how the navigation guard in this file spent its life.
-    expect(localePacks.length, 'no locale packs found in stack.translations').toBeGreaterThan(0);
-    return localePacks;
-  };
-
-  it('every authored field label and help string is translated', () => {
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const obj of objects) {
-        const fields = pack.objects?.[obj.name]?.fields ?? {};
-        for (const [name, f] of Object.entries<AnyRec>(obj.fields ?? {})) {
-          // `help` mirrors the authored `description` — the field-level spelling
-          // differs between the two surfaces (spec: FieldTranslationSchema).
-          if (f?.label && !fields[name]?.label) bad.push(`${locale}: ${obj.name}.${name}.label`);
-          if (f?.description && !fields[name]?.help) bad.push(`${locale}: ${obj.name}.${name}.help`);
-        }
-      }
-    }
-    expect(bad, `fields with no translation:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every page has translated nav copy and header copy', () => {
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const page of pages) {
-        const t = pack.pages?.[page.name];
-        if (page.label && !t?.label) bad.push(`${locale}: ${page.name}.label`);
-        if (page.description && !t?.description) bad.push(`${locale}: ${page.name}.description`);
-
-        // A page's header copy is addressed by the PAGE name: `page:header`
-        // instances carry no stable id, so `pages.<name>.title` / `.subtitle` are
-        // the only keys that reach them. `title` legitimately falls back to
-        // `label`, so it is only required when the two differ.
-        const header = (page.regions ?? [])
-          .flatMap((r: AnyRec) => r.components ?? [])
-          .find((c: AnyRec) => c?.type === 'page:header');
-        const title = header?.properties?.title;
-        const subtitle = header?.properties?.subtitle;
-        if (title && title !== page.label && !t?.title) bad.push(`${locale}: ${page.name}.title`);
-        if (subtitle && !t?.subtitle) bad.push(`${locale}: ${page.name}.subtitle`);
-      }
-    }
-    expect(bad, `pages with untranslated copy:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every dashboard widget has a translated title and description', () => {
-    const dashboards: AnyRec[] = (stack as any).dashboards ?? [];
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const d of dashboards) {
-        const widgets = pack.dashboards?.[d.name]?.widgets ?? {};
-        for (const w of d.widgets ?? []) {
-          if (!w?.id) continue;
-          if (w.title && !widgets[w.id]?.title) bad.push(`${locale}: ${d.name}.${w.id}.title`);
-          if (w.description && !widgets[w.id]?.description) {
-            bad.push(`${locale}: ${d.name}.${w.id}.description`);
-          }
-        }
-      }
-    }
-    expect(bad, `widgets with no translated copy:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every view empty state is translated', () => {
-    // The empty state is the ONE string a view shows when it has no rows — the
-    // exact moment a trial user is most likely to be reading it.
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const record of views) {
-        const containers: [string, AnyRec][] = [
-          ...(record.list ? [['list', record.list] as [string, AnyRec]] : []),
-          ...Object.entries<AnyRec>(record.listViews ?? {}),
-        ];
-        for (const [key, view] of containers) {
-          const empty = view?.emptyState;
-          if (!empty?.title && !empty?.message) continue;
-          const objectName = view?.data?.object ?? record.list?.data?.object ?? record.object;
-          const name = view?.name ?? key;
-          const t = pack.objects?.[objectName]?._views?.[name]?.emptyState;
-          if (empty.title && !t?.title) bad.push(`${locale}: ${objectName}.${name}.emptyState.title`);
-          if (empty.message && !t?.message) {
-            bad.push(`${locale}: ${objectName}.${name}.emptyState.message`);
-          }
-        }
-      }
-    }
-    expect(bad, `empty states with no translated copy:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  /**
-   * ── View LABELS, the surface with no gate at all (#783, #767) ────────────
-   *
-   * The empty-state guard above covers the string a view shows when it has NO
-   * rows. Its label — the tab caption a user reads on every visit — had no
-   * guard of any kind, and two separate reviews found gaps by comparing the
-   * four bundles column by column with human eyes.
-   *
-   * Nothing mechanical could have found them:
-   *
-   *  - `pnpm lint` hard-codes `--skip-i18n`. Dropping the flag does not report
-   *    these either — the run emits one line, `platform built-ins: 2298 i18n
-   *    issue(s) hidden`, because app-authored `_views` completeness is not in
-   *    the set the linter checks at all. So this is NOT the #494 family (real
-   *    warnings suppressed by a flag); it is a surface nobody was checking.
-   *  - Every gap was in `en`, and in the source locale a missing key is
-   *    invisible: the resolver falls back to the metadata `label`, which is
-   *    already correct English. #679 wrote this down for select options, page
-   *    copy and empty states — "a missing key and a correct one render
-   *    identically" — and ruled it a defect anyway, because every other bundle
-   *    is authored by mirroring this one's shape: a view with no slot in `en`
-   *    is a view the next translator has nowhere to put.
-   *
-   * The requirement is therefore the same one the select-field guard states,
-   * applied to view labels: every canonical view, every locale, no exemptions.
-   * No ledger is introduced here — the nine rows this guard was born red on
-   * (`crm_task.todays_tasks` / `overdue_tasks` from #783, `crm_lead.hot_leads`
-   * / `crm_account.renewals_due` / `at_risk_accounts` from #767, and the four
-   * `crm_case` / `crm_opportunity` rows the derivation turned up beside them)
-   * were closed in the same PR. An empty ledger is an invitation to add a row.
-   */
-
-  /**
-   * Every canonical list view in the stack, as `{ object, name }`.
-   *
-   * Derived from the compiled stack rather than from a hand-kept list, which
-   * is the whole point: a view added tomorrow is held to the bar on the PR
-   * that adds it, without anyone remembering to extend a fixture. Same
-   * container walk as the empty-state guard above — a view record carries one
-   * default `list` plus a `listViews` map, and both are user-visible tabs.
-   */
-  const canonicalViews = (): Array<{ object: string; name: string; label?: string }> =>
-    views.flatMap((record) => {
-      const containers: [string, AnyRec][] = [
-        ...(record.list ? [['list', record.list] as [string, AnyRec]] : []),
-        ...Object.entries<AnyRec>(record.listViews ?? {}),
-      ];
-      return containers.map(([key, view]) => ({
-        object: view?.data?.object ?? record.list?.data?.object ?? record.object,
-        name: view?.name ?? key,
-        label: view?.label,
-      }));
-    });
-
-  it('sees a non-trivial set of canonical views and a _views table in every locale', () => {
-    // Guards the guard. The three assertions below iterate derived collections,
-    // and all would pass by checking nothing if `stack.views` came back empty, if
-    // the container walk stopped yielding, or if a bundle's `_views` tables
-    // failed to flatten — the exact way the navigation guard in this file spent
-    // its life green.
-    const derived = canonicalViews();
-    expect(derived.length, 'no canonical list views discovered').toBeGreaterThanOrEqual(50);
-    expect(
-      new Set(derived.map((v) => v.object)).size,
-      'canonical views collapsed onto too few objects — derivation is broken',
-    ).toBeGreaterThanOrEqual(10);
-    expect(
-      derived.filter((v) => typeof v.object === 'string' && typeof v.name === 'string').length,
-      'some canonical view resolved no object or no name',
-    ).toBe(derived.length);
-
-    for (const [locale, pack] of packs()) {
-      const tables = Object.values<AnyRec>(pack.objects ?? {}).filter((o) => o?._views);
-      expect(tables.length, `${locale}: no _views table parsed out of the bundle`).toBeGreaterThan(0);
-    }
-  });
-
-  it('every canonical view label is translated in every locale', () => {
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const { object, name } of canonicalViews()) {
-        if (!pack.objects?.[object]?._views?.[name]?.label) {
-          bad.push(`${locale}: ${object}._views.${name}.label`);
-        }
-      }
-    }
-    expect(
-      bad,
-      `canonical views with no translated label:\n  ${bad.join('\n  ')}\n` +
-        'Add the entry to src/translations/<locale>.ts. In en the gap is ' +
-        'invisible — the resolver falls back to the metadata label — which is ' +
-        'why it needs a test rather than a reviewer.',
-    ).toEqual([]);
-  });
-
-  it('no locale carries a _views entry for a view the stack does not ship', () => {
-    // The other direction of the same parity. A renamed or deleted view leaves
-    // four orphan entries behind, and an orphan is worse than a gap: it reads
-    // as coverage while translating nothing, and the next person to add a view
-    // by that name inherits a stale label.
-    const shipped = new Set(canonicalViews().map(({ object, name }) => `${object}.${name}`));
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const [object, entry] of Object.entries<AnyRec>(pack.objects ?? {})) {
-        for (const name of Object.keys(entry?._views ?? {})) {
-          if (!shipped.has(`${object}.${name}`)) bad.push(`${locale}: ${object}._views.${name}`);
-        }
-      }
-    }
-    expect(bad, `_views entries naming no shipped view:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('the en view label is the canonical label verbatim, so a rename cannot pass silently', () => {
-    // What the parity assertion above CANNOT catch, and #767's actual concern:
-    // the key exists in all four bundles and the metadata label is then edited.
-    // Before this pin, `en` tracked the rename for free (its entry is a copy of
-    // the same English string, and where it was missing the resolver fell back),
-    // while zh-CN / ja-JP / es-ES kept translating the OLD name — four bundles
-    // describing two different views, with nothing red.
-    //
-    // Holding `en` byte-identical to the metadata label is what converts that
-    // into a build failure: rename `⏰ Open Tasks · Most Overdue First` and this
-    // goes red, which puts the author in the translations file with all four
-    // locales in front of them. Be honest about the limit — it proves the
-    // ENGLISH pair agrees, and nothing here can verify that a human then
-    // re-translated the other three. It makes the rename loud; it does not make
-    // it correct.
-    //
-    // This is not a new rule so much as the one the bundle already followed:
-    // all 61 `en` view labels that existed when this landed were already
-    // byte-identical to their metadata label, because #679 extracted them
-    // programmatically from `objectstack.config` rather than transcribing them.
-    // If English ever genuinely wants to read differently from the metadata,
-    // the fix is contract-first — correct the `label` in `*.view.ts`, the one
-    // place every locale is derived from — not to let two English strings drift.
-    const en = packFor('en');
-    expect(en, 'no en pack in stack.translations').toBeTruthy();
-    const bad: string[] = [];
-    for (const { object, name, label } of canonicalViews()) {
-      if (typeof label !== 'string') continue;
-      const translated = en?.objects?.[object]?._views?.[name]?.label;
-      if (typeof translated === 'string' && translated !== label) {
-        bad.push(`${object}.${name}: metadata "${label}" vs en "${translated}"`);
-      }
-    }
-    expect(bad, `en view labels that drifted from the metadata label:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  it('every action parameter label is translated', () => {
-    // A param label is the field caption inside an action's modal — untranslated,
-    // it is a bare English word on an otherwise localized form.
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const action of stackActions) {
-        if (!action.objectName) continue;
-        const params = pack.objects?.[action.objectName]?._actions?.[action.name]?.params ?? {};
-        for (const p of action.params ?? []) {
-          const key = p?.name ?? p?.field;
-          if (key && p?.label && !params[key]?.label) {
-            bad.push(`${locale}: ${action.objectName}.${action.name}.params.${key}.label`);
-          }
-        }
-      }
-    }
-    expect(bad, `action params with no translated label:\n  ${bad.join('\n  ')}`).toEqual([]);
-  });
-
-  /**
-   * Section headings — the surface `objectstack lint` cannot see (#683).
-   *
-   * A section is named in two independent places, and both must be collected or
-   * the requirement is understated:
-   *
-   *  1. Every distinct `group` on an object's fields. `fieldGroups` declares the
-   *     English heading for each, but a group used by a field and absent from
-   *     `fieldGroups` still renders a heading, so the FIELDS are the authority
-   *     for which sections exist.
-   *  2. Every `sections[].name` in the page and view tree — at ANY depth. Detail
-   *     pages nest them inside `page:tabs` → components → `properties.sections`,
-   *     and a shallow walk silently misses exactly those, which is how a first
-   *     pass at this derivation under-counted by nine.
-   */
-  const sectionsByObject = (): Map<string, Set<string>> => {
-    const need = new Map<string, Set<string>>();
-    const add = (objectName?: string, name?: string) => {
-      if (!objectName || !name) return;
-      if (!need.has(objectName)) need.set(objectName, new Set());
-      need.get(objectName)!.add(name);
-    };
-
-    for (const obj of objects) {
-      for (const f of Object.values<AnyRec>(obj.fields ?? {})) add(obj.name, f?.group);
-    }
-
-    const walk = (node: AnyRec | AnyRec[] | undefined, bound?: string): void => {
-      if (!node || typeof node !== 'object') return;
-      if (Array.isArray(node)) { for (const x of node) walk(x, bound); return; }
-      const here = node.object ?? node.properties?.object ?? node.data?.object ?? bound;
-      for (const arr of [node.sections, node.properties?.sections]) {
-        if (Array.isArray(arr)) for (const s of arr) add(here, s?.name);
-      }
-      // `fields` is skipped: it is a map of field definitions, not layout, and
-      // descending into it would attribute a field named `sections` as layout.
-      for (const [key, value] of Object.entries(node)) {
-        if (key !== 'fields' && value && typeof value === 'object') walk(value as AnyRec, here);
-      }
-    };
-    for (const page of pages) walk(page, page.object);
-    for (const view of views) walk(view, view.object ?? view.list?.data?.object);
-
-    return need;
-  };
-
-  it('sees a non-trivial set of sections to check', () => {
-    // Without this, a derivation that silently returned nothing would make the
-    // assertion below pass while checking nothing — the failure mode this whole
-    // describe block exists to rule out.
-    const need = sectionsByObject();
-    const total = [...need.values()].reduce((n, s) => n + s.size, 0);
-    expect(need.size, 'no objects declare sections — derivation is broken').toBeGreaterThan(10);
-    expect(total, 'suspiciously few sections found — derivation is broken').toBeGreaterThan(60);
-  });
-
-  it('every form and detail-page section heading is translated', () => {
-    const need = sectionsByObject();
-    const bad: string[] = [];
-    for (const [locale, pack] of packs()) {
-      for (const [objectName, names] of need) {
-        const translated = pack.objects?.[objectName]?._sections ?? {};
-        for (const name of names) {
-          if (!translated[name]?.label) bad.push(`${locale}: ${objectName}.${name}`);
-        }
-      }
-    }
-    expect(bad, `section headings with no translation:\n  ${bad.join('\n  ')}`).toEqual([]);
   });
 });
